+++ config/commands.php
@@ -1,6 +1,5 @@
 <?php
 
-use App\Commands\Get\BaseGetCommand;
 use Illuminate\Console\Scheduling\ScheduleFinishCommand;
 use Illuminate\Console\Scheduling\ScheduleListCommand;
 use Illuminate\Console\Scheduling\ScheduleRunCommand;
@@ -67,7 +66,6 @@ return [
     */
 
     'hidden' => Phar::running() ? [
-        BaseGetCommand::class,
         DataMakeCommand::class,
         DumpCompletionCommand::class,
         HelpCommand::class,
@@ -86,7 +84,6 @@ return [
     */
 
     'remove' => Phar::running() ? [
-        BaseGetCommand::class,
         DataStructuresCacheCommand::class,
         ScheduleFinishCommand::class,
         ScheduleListCommand::class,